requirements.txt
fastapi==0.110.0
uvicorn==0.21.1
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
SQLAlchemy==1.4.9
alembic==1.8.0  # Optional for database migrations
aiomysql==0.0.21  # Optional for async MySQL support
python-jose[cryptography]==3.3.0  # Optional for JWT authentication
celery==5.2.0  # Optional for background tasks
pytest==7.1.2  # Optional for testing
requests==2.27.1  # Optional for making HTTP requests in tests
pytest-asyncio==0.15.1  # Optional for async testing support with pytest
sphinx==4.3.0  # Optional for documentation generation
sphinx-autoapi==1.3.0  # Optional for auto-generating API documentation
black==22.1.0  # Optional for code formatting
flake8==4.0.1  # Optional for code linting
psycopg2-binary==2.9.3
python-dotenv>=0.21.0
itsdangerous==2.0.1
gunicorn